Promoting shared examples to httptest
