+++ playground/routes_test.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 
 	"github.com/kkrull/gohttp/http"
-	"github.com/kkrull/gohttp/httptest"
 	"github.com/kkrull/gohttp/playground"
+	"github.com/kkrull/gohttp/httptest"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -53,15 +53,9 @@ var _ = Describe("ReadOnlyRoute", func() {
 					routedRequest.Handle(response)
 				})
 
-				It("responds 200 OK with no body", ShouldHaveNoBody(response, 200, "OK"))
-				It("sets Allow to the methods implemented by this type", func() {
-					responseMessage := httptest.ParseResponse(response)
-					responseMessage.HeaderShould("Allow", ContainSubstrings([]string{
-						"GET",
-						"HEAD",
-						"OPTIONS",
-					}))
-				})
+				It("responds 200 OK with no body", httptest.ShouldHaveNoBody(response, 200, "OK"))
+				It("sets Allow to the methods implemented by this type",
+					httptest.ShouldAllowMethods(response, "GET", "HEAD", "OPTIONS"))
 			})
 
 			It("returns nil for any other method", func() {
@@ -122,17 +116,9 @@ var _ = Describe("ReadWriteRoute", func() {
 					routedRequest.Handle(response)
 				})
 
-				It("responds 200 OK with no body", ShouldHaveNoBody(response, 200, "OK"))
-				It("sets Allow to the methods implemented by this type", func() {
-					responseMessage := httptest.ParseResponse(response)
-					responseMessage.HeaderShould("Allow", ContainSubstrings([]string{
-						"GET",
-						"HEAD",
-						"OPTIONS",
-						"POST",
-						"PUT",
-					}))
-				})
+				It("responds 200 OK with no body", httptest.ShouldHaveNoBody(response, 200, "OK"))
+				It("sets Allow to the methods implemented by this type",
+					httptest.ShouldAllowMethods(response, "GET", "HEAD", "OPTIONS", "POST", "PUT"))
 			})
 
 			It("routes POST to Resource#Post", func() {
